+++ test/deprecated.jl
@@ -1,4 +1,3 @@
-# ¡ joinpath does not give deprecation warning, but uses URIs' joinpath instead
 @testset "Deprecations" begin
     @test (@test_deprecated SFTP.pwd(sftp.uri)) == "/pub/example/"
     @test (@test_deprecated SFTP.splitdir(sftp.uri)) == (URI(sftp.uri, path="/pub/"), "example")